Revert "Streamlined Code from Brent"
This reverts commit 88159a0f67f1e2fcef8d272a88abf04db65dd897.
