+++ controllers/uploadController.js
@@ -6,7 +6,6 @@ const crypto = require('crypto');
 const path = require('path');
 //const filetype = require('file-type');
 
-
 // Set up Multer middleware to handle file uploads
 const upload = multer().single('uploadedImage');
 
@@ -114,28 +113,4 @@ const addProfilePicture = asyncHandler(async (req, res) => {
   }
 });
 
-const getProfilePicture = asyncHandler(async (req, res) => { 
-  const _imageID = req.body.imageID;
-
-  if (!_imageID) {
-    return res.status(400).json({ 'message': 'Image ID is required.' });
-  }
-
-  // Verify image ID is a string of 24 hex characters
-    if (!/^[0-9a-fA-F]{24}$/.test(_imageID)) {
-        return res.status(400).json({ 'message': 'Image ID is invalid.' });
-    }
-
-  const imageFound = await Image.findOne({ _id: new ObjectId(_imageID) });
-
-  if (imageFound) {
-    // Retrieve the URL
-    const _imageURL = imageFound.imageURL;
-    console.log(_imageURL); 
-    return res.status(200).json({ 'success': _imageURL });
-  } else {
-    return res.status(400).json({ 'message': 'Image ID not found.' });
-  }
-});
-
-module.exports = { addProfilePicture, getProfilePicture , upload };
+module.exports = { addProfilePicture, upload };
